new branch with 3 new test cases for TrainerTest service

+++ src/main/java/com/revature/assignforce/controllers/TrainerController.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
+//import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -69,7 +69,10 @@ public class TrainerController {
 	// delete
 	@DeleteMapping(value = "{id}")
 	public ResponseEntity<Trainer> delete(@PathVariable int id) {
+		if (trainerService.findById(id) == null)
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		trainerService.delete(id);
 		return new ResponseEntity<>(HttpStatus.OK);
-	}
+ 	}
+	
 }
